+++ packages/webblocks-cms/routes/auth.php
@@ -2,10 +2,24 @@
 
 use Illuminate\Support\Facades\Route;
 use WebBlocks\Cms\Http\Controllers\Auth\LoginController;
+use WebBlocks\Cms\Http\Controllers\Auth\NewPasswordController;
+use WebBlocks\Cms\Http\Controllers\Auth\PasswordResetLinkController;
 
 Route::middleware(['web', 'guest'])->group(function () {
   Route::get('/webadmin/login', [LoginController::class, 'create'])->name('login');
   Route::post('/webadmin/login', [LoginController::class, 'store']);
+
+  Route::get('/webadmin/forgot-password', [PasswordResetLinkController::class, 'create'])
+    ->name('webblocks.auth.password.request');
+
+  Route::post('/webadmin/forgot-password', [PasswordResetLinkController::class, 'store'])
+    ->name('webblocks.auth.password.email');
+
+  Route::get('/webadmin/reset-password/{token}', [NewPasswordController::class, 'create'])
+    ->name('webblocks.auth.password.reset');
+
+  Route::post('/webadmin/reset-password', [NewPasswordController::class, 'store'])
+    ->name('webblocks.auth.password.store');
 });
 
 Route::middleware(['web', 'auth'])->group(function () {